LayoutTests/platform/ios-sim-deprecated/fast/forms/color/input-color-onchange-event-expected.txt

CONSOLE MESSAGE: line 32: TypeError: 'undefined' is not a function (evaluating 'internals.selectColorInColorChooser(input, '#ff0000')')
Test if change event fires properly when color chooser changes. Bug 66848 
To manually test this, click on the input color element in the top left corner and change the value from the color chooser. See if the number of "value changed" messages matches the number of times you changed the color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
